Initial Commit
It's pretty much working. I've verified that the enable/disable feature works, and that it still works even when automatic transmission is selected (which is notable because I'm copying the gear indicator to make the score display). I'm not happy with the centering of the display on the screen, but I can't figure out how Unity objects work to get the size of the thing, so I gave up and just subtracted 50 pixels. It's still a little to the right on my ultra wide, but close enough.